Core: Improve gathering partitions to be expired
When formulating the SQL statement to gather relevant partitions to be
expired, make sure they are processed in order.

The `ORDER BY` is used to process partitions in chronological order, to
not produce temporary gaps. Imagine a table is partitioned by month: If
partitions which are selected for expiry are not processed in order, the
recipe could first drop February, then March, and then January, which
would produce inconsistent gaps in the data, if it was queried at the
same time concurrently, while the expiration process is taking place.