CI: Fix MongoDB tests
`testcontainers-mongodb` needs to be pulled into a separate extra,
because it would otherwise pull in `pymongo` all the time, which in turn
would signal corresponding tests to be invoked.

However, the MongoDB service is not available on all CI workflows, so
things go south.

How could this ever have worked?