API design: Unify input parameters on entry points across the board
The current set of input parameters for addressing a database cluster,
`cluster_id`, `cluster_name`, `sqlalchemy_url`, and `http_url`, needed
a few bits of bundling, so they are now backed by a real model class.

Likewise, the `ManagedCluster` and `StandaloneCluster` classes now
provide the `adapter` property, which will return a suitable
`DatabaseAdapter` instance. This significantly connects the dots for a
better composite object model.

In this spirit, the `DatabaseCluster` class emerges as a new entry point
to acquire a database cluster handle universally.